backend/requirements.txt
fastapi>=0.68.0,<0.69.0
uvicorn>=0.15.0,<0.16.0
python-multipart>=0.0.5,<0.1.0
pandas>=1.3.0,<1.4.0
numpy>=1.20.0,<1.21.0
scipy>=1.7.0,<1.8.0
statsmodels>=0.13.0,<0.14.0
celery>=5.1.0,<5.2.0
redis>=4.1.0,<4.2.0
python-jose[cryptography]>=3.3.0,<3.4.0
pydantic>=1.9.0,<1.10.0
python-dotenv>=0.19.0,<0.20.0
gunicorn==21.2.0
